move the navigation related code to navigation folder
