+++ packages/stacked_services/lib/src/navigation/route_observer.dart
